pyproject.toml
[project]
name = "instaboxd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.11",
    "pillow>=11.1.0",
    "pilmoji>=2.0.2",
    "requests>=2.32.3",
]
